requirements.txt
PySide6==6.9.0
pyserial==3.5
cantools==39.4.2
python-can==4.4.2